pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokerena"
version = "0.1.0"
description = "Pokemon Showdown harness scaffold for local LLM servers."
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
pokerena = "pokerena.cli:main_entry"

[tool.setuptools.packages.find]
where = ["src"]

